segmentation: switch to 3D adapted Sauvola
- 2_1: call savola_3D_image with threeD=True and an anisotropic 3D window
  (9, 21, 21) [(z,y,x); z ~ 21*0.325/0.766 keeps it ~isotropic]; save the result
  as binary_sauvola3d_w9_21_21_k0_15.tiff.
- src/threshold.py: adapted_sauvola_threshold_3d casts to float32 (not float64)
  so the whole-stack 3D uniform filters fit in memory.
- point the downstream loads (2_3, 3, 4) at the new 3D segmentation filename.

+++ src/threshold.py
@@ -68,7 +68,7 @@ def adapted_sauvola_threshold_3d(image, window_size=(5, 15, 15), k=0.1, r=None):
     threshold : 3D ndarray
     binary    : 3D boolean array
     """
-    image = image.astype(np.float64)
+    image = image.astype(np.float32)  # float32 (not 64) to keep the full-stack 3D filters in memory
 
     if r is None:
         r = 0.5 if image.max() <= 1.0 else 128.0
